+++ src/main/resources/static/styles/create-slideshow.css
