+++ src/main/resources/templates/admin/admin.html
@@ -23,9 +23,9 @@
         <h1>Admin Page</h1>
 
         <div class="index-buttons-container">
-            <a href="/admin/manage"><button class="btn dots-button">Manage Images</button></a>
-            <a href="/admin/select-images"><button class="btn sawtooth-button">Create A New Slideshow</button></a>
-            <a href="/admin/slideshows/select"><button class="btn sawtooth-button">Manage Slideshows</button></a>
+            <a href="/admin/images"><button class="btn dots-button">Manage Images</button></a>
+            <a href="/admin/slideshows/create"><button class="btn sawtooth-button">Create A New Slideshow</button></a>
+            <a href="/admin/slideshows"><button class="btn sawtooth-button">Manage Slideshows</button></a>
         </div>
 
     </div>
