+++ src/main/resources/templates/admin/images/edit-image.html
@@ -25,7 +25,7 @@
     <div class="imagedetails">
         <img th:src="${image.url}">
         <div>
-            <form class="formdetails" th:action="@{/admin/images/image-edit/edited}" method="post">
+            <form class="formdetails" th:action="@{/admin/images/edit}" method="post">
                 <label for="id">
                     <span>ID</span>
                     <input id="id" name="id" th:value="${image.id}" readonly/>
