--- src/main/resources/templates/admin/images/edit-images.html
@@ -1,1 +1,0 @@
-<!--TODO Move `edit.html` here-->
