+++ src/main/resources/templates/admin/slideshow/edit-slideshow.html
@@ -1,1 +1,125 @@
-<!--TODO Move `edit-slideshow.html` here-->
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+
+<!--@thymesVar id="model" type="org.softwire.training.slideshowbob.models.pages.EditSlideshowPageModel"-->
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/css/bootstrap.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/js/bootstrap.min.js"></script>
+    <title>Group Project - Home</title>
+
+    <link href="https://fonts.googleapis.com/css?family=Arima+Madurai:200" rel="stylesheet">
+    <link rel="stylesheet" href="/styles/general.css"/>
+    <link rel="stylesheet" href="/styles/navbar.css"/>
+    <link rel="stylesheet" href="/styles/uploadImage.css"/>
+    <link rel="stylesheet" href="/styles/create-slideshow.css"/>
+
+    <script src="https://code.jquery.com/jquery-3.4.0.min.js"></script>
+    <script src="/js/create-slideshow.js"></script>
+    <script src="/js/edit-slideshow.js"></script>
+
+</head>
+<body>
+<div th:replace="fragments/navbar :: navbar">
+    <!-- This element is replaced by the HTML in fragments/navbar.html -->
+</div>
+
+<div class="main-content">
+    <h1 th:text="'Slideshow ' + ${model.getSlideshow().getSlideshowName()}" ></h1>
+    <div class="thumbnail-image-div">
+        <a type="button" class="thumbnail-image" data-toggle="modal" data-target="#modal2">
+            <img class="thumbnail-image" alt="upload icon"
+                 src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTbMnLmCcifVdReL3JQny4hNf-sAr-IpiSnFA4jmraUCqeV4v07"/>
+            <h4>Add Existing Image</h4>
+        </a>
+    </div>
+    <div class="thumbnail-image-div">
+        <a type="button" class="thumbnail-image" data-toggle="modal" data-target="#myModal">
+            <img class="thumbnail-image" alt="upload icon"
+                 src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTbMnLmCcifVdReL3JQny4hNf-sAr-IpiSnFA4jmraUCqeV4v07"/>
+            <h4>Upload Image</h4>
+        </a>
+    </div>
+    <div class="thumbnails">
+        <div class="thumbnail-image-div" th:each="slide:${model.slides}">
+            <a><img class="thumbnail-image" th:src="${slide.slide.url}"/></a>
+            <div class="overlay">
+                <form th:action="'/edit/' + ${slide.slideshowId} + '/slide-delete/' + ${slide.slideId}" method="post">
+                    <button class="imagebuttons" type="submit">Delete</button>
+                </form>
+            </div>
+        </div>
+    </div>
+
+    <!-- Modal -->
+    <div id="myModal" class="modal fade" role="dialog">
+        <div class="modal-dialog">
+
+            <!-- Modal content-->
+            <div class="modal-content">
+                <div class="modal-header">
+                    <button type="button" class="close" data-dismiss="modal">&times;</button>
+                    <h4 class="modal-title">Upload an Image:</h4>
+                </div>
+
+                <div class="modal-body">
+                    <form th:action="'/edit/' + ${model.slideshow.id} + '/add'" method="post">
+                        <label class="modal-form-field">
+                            <span>Author: </span>
+                            <input id="author" type="text" name="author" required>
+                        </label>
+                        <label class="modal-form-field">
+                            <span>Image Name: </span>
+                            <input id="imageName" type="text" name="imageName" required>
+                        </label>
+                        <label class="modal-form-field">
+                            <span>Licence: </span>
+                            <input id="license" type="text" name="license" required>
+                        </label>
+                        <label class="modal-form-field">
+                            <span>Image URL: </span>
+                            <input id="url" type="text" name="url" required>
+                        </label>
+                        <span><!-- spacer --></span>
+                        <input class="center-button" type="submit" value="Upload Image"/>
+                    </form>
+                </div>
+            </div>
+
+        </div>
+    </div>
+
+    <div id="modal2" class="modal fade" role="dialog">
+        <div class="modal-dialog">
+
+            <!-- Modal content-->
+            <div class="modal-content">
+                <div class="modal-header">
+                    <button type="button" class="close" data-dismiss="modal">&times;</button>
+                    <h4 class="modal-title">Upload an Existing Image:</h4>
+                </div>
+
+                <div class="modal-body">
+                    <div class="custom-select" style="width:200px;">
+                        <select>
+                            <option value="0">Select an image:</option>
+                            <option th:each="image:${model.allImages}"
+                                    data-th-attr="data-url=${image.url}, data-name=${image.imageName}, data-id=${image.id}"
+                                    th:text="${image.imageName}"></option>
+                        </select>
+                        <div class="selected-images-list">
+                            <button type="submit" class="buttonStyle" onclick="addExistingItemToSlideshowList()">Add</button>
+                            <ol id="selections"></ol>
+                            <button class="buttonStyle" onclick="updateSlideshow()">Update</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+
+        </div>
+    </div>
+</div>
+</body>
