+++ src/main/resources/templates/fragments/navbar.html
@@ -12,10 +12,10 @@
                 <a href="/admin">Admin Home</a>
             </div>
             <div class="navbar-item">
-                <a href="/admin/manage">Manage Images</a>
+                <a href="/admin/images">Manage Images</a>
             </div>
             <div class="navbar-item">
-                <a href="/admin/slideshows/select">Manage Slideshows</a>
+                <a href="/admin/slideshows">Manage Slideshows</a>
             </div>
             <div class="navbar-item">
                 <a href="/slideshow">Play Slideshow</a>
